+++ src/v2/mod.rs
@@ -949,6 +949,14 @@ impl IntoRequest<generated::PoolInfoRequest> for (&BlockIdentifier, types::Baker
     }
 }
 
+impl IntoRequest<generated::BakerId> for types::BakerId {
+    fn into_request(self) -> tonic::Request<generated::BakerId> {
+        tonic::Request::new(generated::BakerId {
+            value: self.id.index,
+        })
+    }
+}
+
 impl IntoRequest<generated::BlocksAtHeightRequest> for &endpoints::BlocksAtHeightInput {
     fn into_request(self) -> tonic::Request<generated::BlocksAtHeightRequest> {
         tonic::Request::new(self.into())
@@ -2166,6 +2174,22 @@ impl Client {
             .await?;
         let node_info = types::NodeInfo::try_from(response.into_inner())?;
         Ok(node_info)
+    }
+
+    /// Get the projected earliest time a baker wins the opportunity to bake a
+    /// block.
+    /// If the baker is not a baker for the current reward period then then the
+    /// timestamp returned is the projected time of the first block of the
+    /// new reward period.
+    /// Note that the endpoint is only available on a node running at least
+    /// protocol version 6.
+    pub async fn get_baker_earliest_win_time(
+        &mut self,
+        bid: types::BakerId,
+    ) -> endpoints::RPCResult<chrono::DateTime<chrono::Utc>> {
+        let ts = self.client.get_baker_earliest_win_time(bid).await?;
+        let local_time = ts.into_inner().try_into()?;
+        Ok(local_time)
     }
 
     /// Get the transaction events in a given block. If the block does not exist