+++ CHANGELOG.md
@@ -3,7 +3,8 @@
   of the baker for the reward period. Requires a node version at least 6.1.
 - Add support for `GetWinningBakersEpoch`. Requires a node version at least 6.1.
 - Add Support for `GetFirstBlockEpoch`. Requires a node version at least 6.1.
-- Add support for `GetBakersRewardPeriod` endpoint. Requires a node version at least 6.1.
+- Add support for `GetBakersRewardPeriod` endpoint. Requires a node with version at least 6.1.
+- Add Support for `GetBakerEarliestWinTime` endpoint. Requires a node with version at least 6.1.
 
 ## 3.0.1